refactor(dao): move the tests into a dao package
